Release 0.1 - Implementado o sistema de funcionários no front e todo o backend

+++ backend/src/service/UserService/DeleteUserService/index.ts
@@ -1,18 +1,24 @@
-import { Request, Response } from "express";
-import { PrismaClient } from "@prisma/client";
+import { Request, Response } from 'express';
+import { PrismaClient } from '@prisma/client';
 
 const prisma = new PrismaClient();
 
 export class DeleteUserService {
-  async execute( req: Request, res: Response) {
-    const { id } = req.params;
+    async execute(req: Request, res: Response) {
+        const { id } = req.params;
+        if (!id) {
+            return res.status(400).json({ Message: 'Id não informado' });
+        }
+        try {
+            await prisma.user.delete({
+                where: {
+                    id: Number(id),
+                },
+            });
 
-    await prisma.user.delete({
-      where: {
-        id: Number(id),
-      },
-    });
-
-    return res.json({ Message: "User deleted successfully" });
-  }
+            return res.status(200).json({ Message: 'Usuário deletado com sucesso' });
+        } catch (error: any) {
+            return res.status(500).json({ message: error.message });
+        }
+    }
 }
